requirements.txt
numpy
pandas
keras
xlrd
openpyxl
h5py
biopython
scikit-learn
scipy
tensorflow 